Version final Entrega 4. Bug fixes y todas las funcionalidades requeridas

+++ app/src/main/java/com/dam/planetstarwars/data/model/Film.kt
@@ -9,7 +9,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Film(
     @PrimaryKey(autoGenerate = true)
-    val filmId: Int = 0,
+    val filmId: Long = 0,
     val title: String,
     val episodeId: Int,
     val director: String,
